+++ tutorials/conversion_examples/example_commonroad_to_sumo.py
@@ -1,19 +1,18 @@
 import os
-import matplotlib.pyplot as plt
 import numpy as np
 from lxml import etree
 import uuid
 
 from commonroad.common.file_reader import CommonRoadFileReader
 from commonroad.common.file_writer import CommonRoadFileWriter, OverwriteExistingFile
-from commonroad.visualization.draw_dispatch_cr import draw_object
+from commonroad.visualization.mp_renderer import MPRenderer
 
-from crdesigner.io.api import commonroad_to_sumo
+from crdesigner.input_output.api import commonroad_to_sumo
 
-from crdesigner.io.gui.toolboxes.gui_sumo_simulation import SUMO_AVAILABLE
+from crdesigner.input_output.gui.toolboxes.gui_sumo_simulation import SUMO_AVAILABLE
 if SUMO_AVAILABLE:
-    from crdesigner.conversion.sumo_map.config import SumoConfig
-    from crdesigner.conversion.sumo_map.cr2sumo.converter import CR2SumoMapConverter
+    from crdesigner.map_conversion.sumo_map.config import SumoConfig
+    from crdesigner.map_conversion.sumo_map.cr2sumo.converter import CR2SumoMapConverter
     from sumocr.interface.sumo_simulation import SumoSimulation
     from sumocr.visualization.video import create_video
 
@@ -39,8 +38,8 @@ except etree.XMLSyntaxError as xml_error:
 if SUMO_AVAILABLE:
     config = SumoConfig.from_scenario_name(str(uuid.uuid4()))
     config.scenario_name = scenario_name
-    converter = CR2SumoMapConverter(scenario.lanelet_network, config)
-    converter.convert_to_net_file(output_folder)
+    converter = CR2SumoMapConverter(scenario, config)
+    converter.create_sumo_files(output_folder)
 
 # -------------------- Option 3: SUMO conversion APIs with Traffic Simulation and Video Creation -----------------------
 
@@ -54,19 +53,17 @@ planning_problem.translate_rotate(-centroid, 0)
 config = SumoConfig.from_scenario_name(scenario_name)
 
 # convert CR to sumo net
-wrapper = CR2SumoMapConverter(scenario.lanelet_network, config)
-wrapper.convert_to_net_file(output_folder)
+wrapper = CR2SumoMapConverter(scenario, config)
+wrapper.create_sumo_files(output_folder)
 tls_lanelet_id = 43513
 traffic_light_system_generated = wrapper.auto_generate_traffic_light_system(tls_lanelet_id)
 
 print(f"Generated Traffic Light System at {tls_lanelet_id}, {traffic_light_system_generated}")
 
 # draw scenario after traffic light generation
-plt.figure(figsize=(25, 25))
-draw_object(wrapper.lanelet_network)
-plt.axis('equal')
-plt.autoscale()
-plt.show()
+rnd = MPRenderer()
+wrapper.lanelet_network.draw(rnd)
+rnd.render(show=True)
 
 # write generated traffic lights back to commonroad file
 scenario.lanelet_network = wrapper.lanelet_network
@@ -94,4 +91,5 @@ CommonRoadFileWriter(simulated_scenario,
     overwrite_existing_file=OverwriteExistingFile.ALWAYS)
 
 print("creating video (this may take some time)")
-create_video(simulation, 1, config.simulation_steps, output_folder)
+create_video(simulation.commonroad_scenarios_all_time_steps(),
+             output_folder, trajectory_pred=simulation.ego_vehicles)