+++ tutorials/conversion_examples/example_opendrive_to_commonroad.py
@@ -5,13 +5,13 @@ from commonroad.scenario.scenario import Tag
 from commonroad.common.file_writer import CommonRoadFileWriter, OverwriteExistingFile
 from commonroad.planning.planning_problem import PlanningProblemSet
 
-from crdesigner.conversion.opendrive.opendriveparser.parser import parse_opendrive
-from crdesigner.conversion.opendrive.opendriveconversion.network import Network
+from crdesigner.map_conversion.opendrive.opendrive_parser.parser import parse_opendrive
+from crdesigner.map_conversion.opendrive.opendrive_conversion.network import Network
 
-from crdesigner.io.api import opendrive_to_commonroad
+from crdesigner.input_output.api import opendrive_to_commonroad
 
 
-input_path = ""  # replace empty string
+input_path = "../../tests/opendrive_test_files/el-vendrell-02.06.20.xodr"
 
 # ----------------------------------------------- Option 1: General API ------------------------------------------------
 # load OpenDRIVE file, parse it, and convert it to a CommonRoad scenario