+++ CHANGELOG.md
@@ -1,6 +1,20 @@
-# Changelog commonroad-map-and-scenario-designer
+# Changelog
 
-## [0.0.5] - 2021-04-01
+## [0.2] - 2021-09-22
+
 ### Added
-- Graphical User Interface for creating and manipulating CommonRoad maps and scenarios
+- Planning problem visualization in GUI
+
+### Changed
+- Extension of readme and documentation
+
+### Fixed
+- Various small bug fixes
+
+## [0.1] - 2021-04-01
+### Added
+- graphical user interface for creating and manipulating CommonRoad maps and scenarios
 - converts between different map formats: OpenStreetMap, SUMO, Lanelet/Lanelet2, OpenDRIVE, CommonRoad
+- test cases for different converters
+- tutorials for different converters
+- readme and documentation